firmware: Refactor USB sending task
Create an async version of send_device_message() to allow cleaning up
the USB-sending implementation quite a bit.

It is important to note that the `RefCell` borrows MUST NEVER be live
over an await.  To archieve this, the `###_maybe` temporary variables
are bound before the `if {}` block so the `Ref`s are dropped before
entering the block where the USB send function is awaited.