host-daemon: Allow multiple independent sets of property matches
There are situations where entirely independent applications with
entirely different property sets should be attached to the same channel.
For example, I like routing all games to channel 4.

To allow this, change the configuration format to keep an array of
property matches instead of just a single one.  If any of the match sets
matches (as implemented before), the sink-input will be attached.